LOD Plot

Creates a LOD plot for a specified 'Target'. The LOD plot shows the relative detection rates for each standard as well as the LOD model curve. This code has been modified and functionalised from the original by Merkes et al. (see references).

calib_plot_lod(lod_obj, target, legend = TRUE, ...)

Arguments

lod_obj
A lod object containing the output from the calib_lod function.

target
A character string containing a unique identifier for the target calibration curve to be plotted.

legend
A logical indicating whether a lengen should be included on the plot, Default: TRUE.

...
Placeholder for further arguments that might be needed by future implementations.

Value

A LOD plot in the current plotting device and a data.frame containing the effective LOD estimates, standard errors, upper and lower confidence intervals for each replicate.

Details

The LOD plot shows the relative detection rates for each standard as well as the LOD model curve. The 95 multiple replicates are displayed. The logarithmic function that was used to determine the LODs is shown in the plot subtitle along with the p‐value for a lack of fit test on the model.

References

Merkes CM, Klymus KE, Allison MJ, Goldberg C, Helbing CC, Hunter ME, Jackson CA, Lance RF, Mangan AM, Monroe EM, Piaggio AJ, Stokdyk JP, Wilson CC, Richter C. (2019) Generic qPCR Limit of Detection (LOD) / Limit of Quantification (LOQ) calculator. R Script. Available at: https://github.com/cmerkes/qPCR_LOD_Calc. DOI: https://doi.org/10.5066/P9GT00GB.

Klymus, Katy E., Christopher M. Merkes, Michael J. Allison, Caren S. Goldberg, Caren C. Helbing, Margaret E. Hunter, Craig A. Jackson, et al. ‘Reporting the Limits of Detection and Quantification for Environmental DNA Assays’. Environmental DNA 2, no. 3 (July 2020): 271–82. https://doi.org/10.1002/edn3.29.

Examples

if (FALSE) {
qpcr_lod <- calib_lod(data = calib_data, threshold = 0.35,
              lod.fit = "best", loq.fit = "best")
lod_CI <- calib_plot_lod(lod_obj = qpcr_lod, target = "10720201")
lod_CI
}